refactor connection
