+++ cmb_protocol/connection.py
@@ -1,0 +1,8 @@
+class Connection:
+    def __init__(self, shutdown, spawn, send):
+        self.shutdown = shutdown
+        self.spawn = spawn
+        self.send = send
+
+    async def handle_packet(self, packet):
+        self.shutdown()
